requirements.txt
huggingface-hub
ultralytics
opencv-python-headless
Pillow
toml
argparse